requirements.txt
FLASK==2.2.5
Flask-SQLAlchemy
SQLAlchemy
flask-restful
requests
flask-login
flask-security-too==5.1.2
Werkzeug==2.3.0
bcrypt
celery